requirements.txt
# Web Knowledge Aggregator Agent Dependencies

# Core Framework - Using working versions from previous project
langchain
langchain-openai
langchain-core
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web Search (No API key needed)
duckduckgo-search>=4.1.1

# Web Scraping
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
requests

# Text Processing
python-dotenv>=0.19.0

# Optional UI (for future)
streamlit>=1.29.0

# Utilities
tiktoken>=0.5.2
lxml>=4.9.3